+++ src/main/java/dev/willyelton/crystal_tools/common/network/handler/ToolAttributeHandler.java
@@ -62,7 +62,6 @@ public class ToolAttributeHandler {
 
                 // update the skill points array
                 DataComponents.addValueToArray(heldTool, DataComponents.POINTS_ARRAY, payload.id(), pointsToAdd);
-
             }
         });
     }